requirements.txt
boto3>=1.28.0
langchain>=0.0.267
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
jsonschema>=4.17.3
